+++ apps/web/components/Feature/Auth/Login/LoginForm/index.tsx
@@ -1,8 +1,7 @@
 "use client";
 
-import { FormEvent, useState, useEffect, useRef } from "react";
+import { useState } from "react";
 import { useTranslation } from "react-i18next";
-import { useRouter } from "next/navigation";
 import { EyeClosedIcon, EyeOpenIcon } from "@/assets/icons";
 import InputField from "@/components/UI/InputFields";
 import {
@@ -21,104 +20,21 @@ import logo from "@/assets/icons/Kiibee_logo_mark_black.svg";
 import Image from "@/components/UI/SafeImage";
 import GenericButton from "@/components/UI/GenericButton";
 import { MonoText } from "@/components/UI/Monotext";
-import { useLoginFormSchema } from "@/utils/useLoginFormSchema";
-import type { LoginFormErrors } from "@/utils/authLoginFormSchema";
 import { ALERT } from "@/utils/common";
-import {
-  getPostLoginPath,
-  persistLoginSession,
-  useLogin,
-} from "@/hooks/auth/useLogin";
 import { PATHS } from "@/utils/path";
 import { INPUT_TYPE } from "@/utils/ui";
+import { useLoginForm } from "@/hooks/Auth/useLoginForm";
 
 export default function LoginForm() {
   const { t } = useTranslation();
-  const router = useRouter();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
   const [isPasswordVisible, setIsPasswordVisible] = useState(false);
-  const [remember, setRemember] = useState(false);
-  const [fieldErrors, setFieldErrors] = useState<LoginFormErrors>({});
-  const [formError, setFormError] = useState("");
-  const loginSchema = useLoginFormSchema();
-  const { mutateAsync: login, isPending } = useLogin();
-  const isMounted = useRef(true);
 
-  useEffect(() => {
-    isMounted.current = true;
-    return () => {
-      isMounted.current = false;
-    };
-  }, []);
+  const { state, actions } = useLoginForm();
+  const { email, password, remember, fieldErrors, formError, isPending } =
+    state;
 
-  const handleSubmit = async (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-
-    const parsedValues = loginSchema.safeParse({
-      email,
-      password,
-    });
-
-    if (!parsedValues.success) {
-      const flattenedErrors = parsedValues.error.flatten().fieldErrors;
-      if (isMounted.current) {
-        setFieldErrors({
-          email: flattenedErrors.email?.[0],
-          password: flattenedErrors.password?.[0],
-        });
-        setFormError(t("authForm.errors.fixHighlightedFields"));
-      }
-      return;
-    }
-
-    if (isMounted.current) {
-      setFormError("");
-      setFieldErrors({});
-    }
-
-    try {
-      const response = await login(parsedValues.data);
-
-      if (response.success === false) {
-        if (isMounted.current) {
-          setFormError(response.message || t("authForm.errors.submitFailed"));
-        }
-        return;
-      }
-
-      persistLoginSession(response);
-      router.push(getPostLoginPath(response));
-    } catch (error) {
-      if (isMounted.current) {
-        setFormError(
-          error instanceof Error
-            ? error.message
-            : t("authForm.errors.submitFailed"),
-        );
-      }
-    }
-  };
-
-  const handleEmailChange = (nextValue: string) => {
-    setEmail(nextValue);
-    if (fieldErrors.email || formError) {
-      if (isMounted.current) {
-        setFieldErrors((prev) => ({ ...prev, email: undefined }));
-        setFormError("");
-      }
-    }
-  };
-
-  const handlePasswordChange = (nextValue: string) => {
-    setPassword(nextValue);
-    if (fieldErrors.password || formError) {
-      if (isMounted.current) {
-        setFieldErrors((prev) => ({ ...prev, password: undefined }));
-        setFormError("");
-      }
-    }
-  };
+  const { setRemember, handleEmailChange, handlePasswordChange, handleSubmit } =
+    actions;
 
   return (
     <Wrapper>
